requirements.txt
fastapi
uvicorn
transformers
pillow